services/moisture-sensor/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
pyModbusTCP==0.3.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
python-jose[cryptography]
scikit-learn==1.1.3          # ✅ มี wheel สำหรับ ARM
numpy==1.22.4                # ✅ มี wheel สำหรับ ARM
scipy==1.9.3                # ✅ มี wheel สำหรับ ARM
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
minimalmodbus==2.1.1
pandas==2.2.3                # ✅ ทดสอบแล้วใช้ร่วมกับ numpy==1.24 ได้
